+++ ak_library_management/static/src/js/customer_fetch.js
@@ -1,6 +1,7 @@
 /** @odoo-module **/
 
 import publicWidget from "@web/legacy/js/public/public_widget";
+import { rpc } from "@web/core/network/rpc";
 
 
 publicWidget.registry.ContactControllerPage = publicWidget.Widget.extend({
@@ -8,12 +9,16 @@ publicWidget.registry.ContactControllerPage = publicWidget.Widget.extend({
      events: {
             'click': '_onFetchClick'
         },
-//        ["email", "=", email ],
-//        ["company_id","vat","phone","mobile","website"],
-//    ])
+//   _onFetchClick function call when button is click.
      _onFetchClick: function(event){
-        console.log("customer")
-//        customer = this.pyEnv["res.partner"].search([])
-
+        var email = $('#email').val();
+//      call rpc method and pass data into customer form
+        rpc('/customer/fetch',{'email':email}).then(
+            function(data){
+            $('#name').val(data.name);
+            $('#company').val(data.company);
+            $('#taxID').val(data.vat);
+            $('#phone').val(data.phone);
+            });
      },
 })